# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for managing animals, species, and categories.",
        "title": "Animal API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/animals": {
            "get": {
                "description": "Get a list of all animals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Get all animals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Animal"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new animal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Create a new animal",
                "parameters": [
                    {
                        "description": "Animal to create",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Animal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Animal"
                        }
                    }
                }
            }
        },
        "/animals/{id}": {
            "get": {
                "description": "Get an animal by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Get an animal by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Animal"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an animal by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Update an animal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Animal to update",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Animal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Animal"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an animal by its ID",
                "tags": [
                    "animals"
                ],
                "summary": "Delete an animal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Get a list of all categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Category"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category to create",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "Get a category by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get a category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a category by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Update a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category to update",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Category"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a category by its ID",
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/species": {
            "get": {
                "description": "Get a list of all species",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "species"
                ],
                "summary": "Get all species",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Species"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new species",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "species"
                ],
                "summary": "Create a new species",
                "parameters": [
                    {
                        "description": "Species to create",
                        "name": "species",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Species"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Species"
                        }
                    }
                }
            }
        },
        "/species/name/{species_name}": {
            "get": {
                "description": "Get a species by its name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "species"
                ],
                "summary": "Get a species by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Species Name",
                        "name": "species_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Species"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Animal": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "animal_name": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/main.Location"
                },
                "species_id": {
                    "type": "string"
                }
            }
        },
        "main.Category": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                }
            }
        },
        "main.Location": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "main.Species": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/main.Location"
                },
                "species_name": {
                    "type": "string"
                }
            }
        }
    }
}